// client/src/constant/VersionSettingType.ts
export enum VersionSettingType {
  LEVEL_1_7 = 1,
  LEVEL_8_10 = 2,
  LEVEL_1_7_NO_SKILL = 3,
  LEVEL_8_10_NO_SKILL = 4,
  LEVEL_ALL = 5,
}

export enum FlagSkillValue {
  HAVE_SKILL = 1,
  NO_SKILL = 2,
}

export enum TypeAchievement {
  PERSONAL_17 = 1,
  PERSONAL_810 = 2,
  DEPARTMENT_810 = 3,
}